requirements.txt
streamlit 
pillow
requests
numpy
python-dotenv
cryptography